+++ DevNet-WebAPI/Controllers/CommentsController.cs
@@ -9,6 +9,8 @@ using DevNet_DataAccessLayer.Data;
 using DevNet_DataAccessLayer.Models;
 using DevNet_WebAPI.Infrastructure.DTO;
 using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
+using Microsoft.Extensions.Hosting;
 
 namespace DevNet_WebAPI.Controllers
 {
@@ -30,7 +32,7 @@ namespace DevNet_WebAPI.Controllers
         {
             return await _context.Comments.ToListAsync();
         }
-
+        
         // GET: api/Comments/5
         [HttpGet("{id}")]
         [Authorize]
@@ -48,11 +50,39 @@ namespace DevNet_WebAPI.Controllers
 
         // PUT: api/Comments/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        /*
-         * Comentado por si se agrega la opcion de editar comentarios
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutComment(Guid id, Comment comment)
+        [Authorize]
+        public async Task<IActionResult> EditComment(Guid id, [FromBody] EditCommentDto commentDto)
         {
+            if (id != commentDto.Id)
+            {
+                return BadRequest();
+            }
+
+            // Obtener el UserId del token JWT
+            var userIdFromToken = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            // Verificar que el UserId del token coincida con el UserId del postDto
+            if (userIdFromToken != commentDto.UserId.ToString())
+            {
+                return Unauthorized("El usuario que intenta editar el comentario no coincide con el que ha iniciado sesion.");
+            }
+
+            var comment = await _context.Comments.FindAsync(id);
+            if (comment == null)
+            {
+                return NotFound();
+            }
+
+            if (comment.UserId != commentDto.UserId)
+            {
+                return Unauthorized("No tienes permiso de editar este comentario.");
+            }
+
+            comment.Text = commentDto.Text;
+            
+
+
             if (id != comment.Id)
             {
                 return BadRequest();
@@ -78,7 +108,6 @@ namespace DevNet_WebAPI.Controllers
 
             return NoContent();
         }
-        */
 
         // POST: api/Comments
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
@@ -86,6 +115,14 @@ namespace DevNet_WebAPI.Controllers
         [Authorize]
         public async Task<ActionResult<Comment>> PostComment([FromBody] PostCommentDto commentDto)
         {
+            var userIdFromToken = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            // Verificar que el UserId del token coincida con el UserId del postDto
+            if (userIdFromToken != commentDto.UserId.ToString())
+            {
+                return Unauthorized("El usuario que intenta dejar un nuevo comentario no coincide con el que ha iniciado sesion.");
+            }
+
             Comment comment = new Comment
             {
                 Id = Guid.NewGuid(),
@@ -104,13 +141,32 @@ namespace DevNet_WebAPI.Controllers
         // DELETE: api/Comments/5
         [HttpDelete("{id}")]
         [Authorize]
-        public async Task<IActionResult> DeleteComment(Guid id)
+        public async Task<IActionResult> DeleteComment(Guid id, [FromBody] DeleteCommentDto deleteDto)
         {
+            if (id != deleteDto.Id)
+            {
+                return BadRequest();
+            }
+            
+            var userIdFromToken = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            // Verificar que el UserId del token coincida con el UserId del postDto
+            if (userIdFromToken != deleteDto.UserId.ToString())
+            {
+                return Unauthorized("El usuario que intenta eliminar el comentario no coincide con el que ha iniciado sesion.");
+            }
+
             var comment = await _context.Comments.FindAsync(id);
             if (comment == null)
             {
                 return NotFound();
             }
+
+            if (comment.UserId != deleteDto.UserId)
+            {
+                return Unauthorized("No tienes permiso de borrar este comentario.");
+            }
+
 
             _context.Comments.Remove(comment);
             await _context.SaveChangesAsync();
